pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "repl-nix-neurodiversejobquest"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "openai>=1.58.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary",
    "email-validator",
    "python-dotenv",
    "boto3 (>=1.37.37,<2.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
]

[tool.poetry]
packages = [
    { include = "app.py" },
    { include = "utils.py" },
    { include = "models.py" },
    { include = "main.py" }
]
